Default to all types in one file
The default is now to place all types in a file named as the .stidgen
file but with the .Generated.cs extension.
This default can be changed by specifying the property UseNameAsFileName
to true. In this case the name of the file is the name of the ID Type
with .Generated.cs appended.
The same property as before (FileName) can also be specified to override
everything else and specify an hardcoded location.

Note: Even if the locations are hardcoded multiple types can be
generated in the same file.